Initial Commit v0.02
This is an initial commit for some new features that can be added. This
version is intended to be changed and is not yet ready for release or
update to the Steam Workshop. Added: Support for changing the map based
on players currently on the server, new features that are not yet ready
have been commented out, and comments include features to be added at a
later date and can be viewed now.